backend/requirements.txt
aniso8601==7.0.0
anyio==4.6.2.post1
async-property==0.2.2
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
deprecation==2.1.0
ecdsa==0.19.0
exceptiongroup==1.2.2
Flask==2.2.3
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
Flask-Reuploaded==1.4.0
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==2.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==3.0.1
packaging==24.1
promise==2.3
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==0.24.0
requests==2.25.1
requests-toolbelt==1.0.0
rsa==4.9
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.54
stripe==11.1.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==2.2.3
zipp==3.20.2